requirements.txt
pytest
selenium
jsonschema
pylint
flake8
allure-pytest
allure-python-commons